rename platformer and spline barriers lessons
