+++ _projects/.makeprojects
@@ -11,7 +11,7 @@
 # === NESTED STRUCTURE (organized by category) ===
 # Games
 games/backgrounds
-games/barriers-spline-sorcerers
+games/spline-barriers:dev
 games/characters
 games/collisions-mechanic
 games/cs-pathway:dev
@@ -22,7 +22,7 @@ games/local-storage
 games/multiplayer
 games/ocean
 games/peppa-pig
-games/platformer-gravity-sorcerers
+games/platformer:dev
 games/red-riding
 games/teamtestgame
 games/transitions
