+++ _projects/games/spline-barriers/levels/SpriteSheetCoin.js
